# perf(bls24-315): faster G2 membership test