--- a/internal/generator/ecc/template/tests/point.go.tmpl
+++ b/internal/generator/ecc/template/tests/point.go.tmpl
@@ -150,6 +150,17 @@ func Test{{ $TAffine }}IsOnCurve(t *testing.T) {
 		{{$fuzzer}},
 	))
 
+	properties.Property("[{{ toUpper .Name }}] IsInSubGroup and MulBy subgroup order should be the same", prop.ForAll(
+		func(a {{ .CoordType}}) bool {
+            var op1, op2 {{ $TJacobian }}
+			op1 = fuzzJacobian{{ $TAffine }}(&{{.PointName}}Gen, a)
+            _r := fr.Modulus()
+            op2.ScalarMultiplication(&op1, _r)
+			return op1.IsInSubGroup() && op2.Z.IsZero()
+		},
+		{{$fuzzer}},
+	))
+
 	properties.TestingRun(t, gopter.ConsoleReporter(false))
 }