@ThomasPiellard 2022-01-04
@yelhousni Can you add a specific test for `IsInSubGroup` ?

Also there is something that bothers me in [https://eprint.iacr.org/2021/1130.pdf](https://eprint.iacr.org/2021/1130.pdf)... The reasoning that I understood from the paper is "psi acts as the multiplication by u on the p+1-t torsion of E, so let's just check that psi(Q)=[u]Q".

 However, nothing says that there is no other torsion subgroup on which psi acts like this. For example, if we find `a` such that 'a | h2', `p^2-1 != 0 [a]` (so E' contains a subgroup of order `a` but not the full `a`-torsion) and `u` is a root of `X^2-tX+p [a]`, then psi acts by multiplication by `u` on the `a` torsion of E'[Fp^2]. From the equations, it's not clear that such a situation does not happen. It seems really unlikely because of the third condition (that `u` has to be a root of `X^2-tX+p [a]`), but unless proceeding case by case there is no guarantee of that.

So I am in favour of keeping the solution before the very end of sect 4, that is checking if psi(uQ)+psi(Q)+..=uQ. The delta that I have between the benches is 
```
benchmark                         old ns/op     new ns/op     delta
BenchmarkG2JacIsInSubGroup-12     122793        126198        +2.77%
```